+++ examples/ekp-showcase.el
@@ -21,7 +21,7 @@
 
 ;;; Code:
 
-(require 'ekp)
+(require 'ekp-layout)
 (require 'ekp-buffer)
 
 (defvar ekp-showcase-min-width 140)